part2/requirements.txt
cryptography~=42.0.5
apache-airflow~=2.8.1